+++ src/behaviors/index.js
@@ -1,3 +1,4 @@
 export { default as draggable } from './draggable';
 export { default as droppable } from './droppable';
 export { default as selectable } from './selectable';
+export { default as scrollable } from './scrollable';
