+++ src/components/Elements/Panel.js
@@ -1,5 +1,4 @@
 import React, { Component } from 'react';
-import Scrollable from '../../containers/Elements/Scrollable';
 
 class Panel extends Component {
 
@@ -13,9 +12,7 @@ class Panel extends Component {
       <div className='panel'>
         <div className='panel__heading'><h5>{ title }</h5></div>
         <div className='panel__content'>
-          <Scrollable>
-            { children }
-          </Scrollable>
+          { children }
         </div>
       </div>
     );
